[CH #156398012] Move test file to separate folder
